refactor: change the folder structure of the project
