+++ src/app.module.ts
@@ -6,7 +6,7 @@ import { ConfigModule } from '@nestjs/config';
 import { ArtistModule } from './modules/artist/artist.module';
 import { CustomExceptionFilter } from './common/filters/custom-exception.filter';
 import { APP_FILTER } from '@nestjs/core';
-import { LoggingModule } from './logging/logging.module';
+import { LoggingModule } from './shared/logging/logging.module';
 
 @Module({
   imports: [
