+++ src/main.ts
@@ -3,7 +3,7 @@ import { AppModule } from './app.module';
 import { ConfigService } from '@nestjs/config';
 import { ValidationPipe } from '@nestjs/common';
 import { DatabaseService } from './database/database.service';
-import { LoggingService } from './logging/logging.service';
+import { LoggingService } from './shared/logging/logging.service';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
