+++ src/modules/album/album.controller.ts
@@ -13,7 +13,7 @@ import { AlbumService } from './album.service';
 import { Album } from './entities/album.entity';
 import { CreateAlbumDto } from './dtos/create-album.dto';
 import { UpdateAlbumDto } from './dtos/update-album.dto';
-import { ErrorHandler } from '../../error/error.handler';
+import { ErrorHandler } from '../../shared/error/error.handler';
 
 @Controller('album')
 export class AlbumController {
