+++ src/modules/user/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
-import { UserRepository } from './user.repository';
+import { UserRepository } from '../../database/repositories/user.repository';
 import { UserService } from './user.service';
 import { DatabaseModule } from '../../database/database.module';
-import { ErrorModule } from '../../error/error.module';
+import { ErrorModule } from '../../shared/error/error.module';
 
 @Module({
   imports: [DatabaseModule, ErrorModule],
